+++ src/formatters/table.rs
@@ -40,17 +40,30 @@ fn format_field_expression_value(
     expression: &Expression,
     shape: Shape,
 ) -> Expression {
-    let singleline_value = format_expression(ctx, expression, shape)
-        .update_trailing_trivia(FormatTriviaType::Replace(vec![])); // We will remove all the trivia from this value, and place it after the comma
+    // Remove all trivia from the output expression as it will be moved after the comma
 
-    if trivia_util::can_hang_expression(expression)
-        && shape.take_first_line(&singleline_value).over_budget()
-        || trivia_util::expression_contains_inline_comments(expression)
-    {
-        hang_expression(ctx, expression, shape, Some(1))
-            .update_trailing_trivia(FormatTriviaType::Replace(vec![]))
+    if trivia_util::can_hang_expression(expression) {
+        if trivia_util::expression_contains_inline_comments(expression) {
+            hang_expression(ctx, expression, shape, Some(1))
+                .update_trailing_trivia(FormatTriviaType::Replace(vec![]))
+        } else {
+            let singleline_value = format_expression(ctx, expression, shape)
+                .update_trailing_trivia(FormatTriviaType::Replace(vec![]));
+            let hanging_value = hang_expression(ctx, expression, shape, Some(1))
+                .update_trailing_trivia(FormatTriviaType::Replace(vec![]));
+
+            if shape.test_over_budget(&singleline_value)
+                || format!("{}", hanging_value).lines().count()
+                    < format!("{}", singleline_value).lines().count()
+            {
+                hanging_value
+            } else {
+                singleline_value
+            }
+        }
     } else {
-        singleline_value
+        format_expression(ctx, expression, shape)
+            .update_trailing_trivia(FormatTriviaType::Replace(vec![]))
     }
 }